--- a/src/lib/i18n/locales/en-GB/translation.json
+++ b/src/lib/i18n/locales/en-GB/translation.json
@@ -650,6 +650,7 @@
 	"Llama.cpp": "",
 	"LLMs can make mistakes. Verify important information.": "",
 	"Loading metrics...": "",
+	"Loading more prompts...": "",
 	"Loading...": "",
 	"Local": "",
 	"Local Models": "",