--- a/src/lib/i18n/locales/fr-CA/translation.json
+++ b/src/lib/i18n/locales/fr-CA/translation.json
@@ -650,6 +650,7 @@
 	"Llama.cpp": "Llama.cpp",
 	"LLMs can make mistakes. Verify important information.": "Les GML peuvent faire des erreurs. Vérifiez les informations importantes.",
 	"Loading metrics...": "Chargement des métriques...",
+	"Loading more prompts...": "Chargement de plus d'invites...",
 	"Loading...": "Chargement...",
 	"Local": "Local",
 	"Local Models": "Modèles locaux",